backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic>=2.0,<3.0
python-multipart
bleach
email-validator
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]
prometheus-fastapi-instrumentator

